Added the ZonesBaseWorker class. Completed unit tests for the ZonesBase namespace.

+++ src/NOAA.NET/Zones/Base/ZonesBaseBuilder.cs
@@ -44,5 +44,5 @@ public sealed class ZonesBaseBuilder
     /// <summary>
     /// Gets or sets the Limit property.
     /// </summary>
-    public int? Limit { get; set; } = 50;
+    public ushort? Limit { get; set; } = 50;
 }
